+++ docs/examples/grep_clone/1_grep_clone.py
@@ -1,0 +1,9 @@
+import arc
+
+
+@arc.command()
+def command():
+    print("hello there!")
+
+
+command()